src/main/java/com/calendar/workout/service/category/CategoryService.java
package com.calendar.workout.service.category;

import com.calendar.workout.domain.category.CategoryRepository;
import com.calendar.workout.dto.category.request.CategoryRequest;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@RequiredArgsConstructor
@Service
public class CategoryService {

    private final CategoryRepository categoryRepository;

    @Transactional
    public Long save(CategoryRequest categoryRequest) {
        return categoryRepository.save(categoryRequest.toEntity()).getId();
    }
}
